pyproject.toml
[build-system]
requires = ["setuptools","setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "todo"
description = "Manage todo under home folder"
authors = [
    {name = "AMoment", email="amoment096@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "toml>=0.10.2",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/A-moment096/MyToolBox"
Issues = "https://github.com/A-moment096/MyToolBox/issues"

[project.scripts]
todo = "todo_manager:main"

[tool.setuptools_scm]
write_to = "todo_manager/_version.py"